fix: share_transfer display wrong currency symbo
In multi-companies setting, on amount and rate field of  Share Transfer
doctype are displaying wrong currency symbol, when create a
Share Transfer of others company that has different currency than
the main company. This due to the lack of `options` on those fields.

To fix this, add `options` to `amount` and `rate` fields.